Point now works with double instead of size_t

+++ src/Point.cpp
@@ -2,7 +2,7 @@
 
 #include "Point.h"
 
-Point::Point(const std::string name, const size_t &xPos, const size_t &yPos) :
+Point::Point(const std::string name, const double &xPos, const double &yPos) :
         Shape::Shape(name, ShapeType::POINT),
         _xPos(xPos),
         _yPos(yPos) {
